requirements.txt
pyrebase4
setuptools
python-dotenv